splitting agents, context
