+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/agent/entity/AIAgentStrategy.kt
@@ -1,5 +1,6 @@
 package ai.koog.agents.core.agent.entity
 
+import ai.koog.agents.core.agent.context.AIAgentContext
 import ai.koog.agents.core.agent.context.AIAgentContextBase
 import ai.koog.agents.core.annotation.InternalAgentsApi
 import ai.koog.agents.core.utils.runCatchingCancellable
@@ -7,7 +8,7 @@ import ai.koog.agents.core.utils.runCatchingCancellable
 /**
  * Represents a strategy interface for managing and executing AI agent workflows.
  */
-public interface AIAgentStrategy<Input, Output> {
+public interface AIAgentStrategy<Input, Output, TContext : AIAgentContextBase<*>> {
     /**
      * Name of the agent strategy
      * */
@@ -20,5 +21,5 @@ public interface AIAgentStrategy<Input, Output> {
      * @param input The input object representing the data to be processed by the AI agent.
      * @return The output of the AI agent execution, generated after processing the input.
      */
-    public suspend fun execute(context: AIAgentContextBase<*>, input: Input): Output?
+    public suspend fun execute(context: TContext, input: Input): Output?
 }
