+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/dsl/builder/AIAgentParallelNodesMergeContext.kt
@@ -2,6 +2,7 @@ package ai.koog.agents.core.dsl.builder
 
 import ai.koog.agents.core.agent.config.AIAgentConfigBase
 import ai.koog.agents.core.agent.context.AIAgentContextBase
+import ai.koog.agents.core.agent.context.AIAgentGraphContext
 import ai.koog.agents.core.agent.context.AIAgentLLMContext
 import ai.koog.agents.core.agent.entity.AIAgentStateManager
 import ai.koog.agents.core.agent.entity.AIAgentStorage
@@ -10,6 +11,7 @@ import ai.koog.agents.core.agent.entity.graph.AIAgentGraphStrategy
 import ai.koog.agents.core.annotation.InternalAgentsApi
 import ai.koog.agents.core.environment.AIAgentEnvironment
 import ai.koog.agents.core.feature.AIAgentFeature
+import ai.koog.agents.core.feature.AIAgentGraphPipeline
 import ai.koog.agents.core.feature.AIAgentPipeline
 import ai.koog.prompt.message.Message
 
@@ -26,9 +28,9 @@ import ai.koog.prompt.message.Message
  */
 @OptIn(InternalAgentsApi::class)
 public class AIAgentParallelNodesMergeContext<Input, Output>(
-    private val underlyingContextBase: AIAgentContextBase<AIAgentGraphStrategy<Input, Output>>,
+    private val underlyingContextBase: AIAgentGraphContext,
     public val results: List<ParallelResult<Input, Output>>,
-) : AIAgentContextBase<AIAgentGraphStrategy<Input, Output>> {
+) : AIAgentContextBase<AIAgentGraphPipeline<*, *>> {
     // Delegate all properties to the underlying context
     override val environment: AIAgentEnvironment get() = underlyingContextBase.environment
     override val id: String get() = underlyingContextBase.id
@@ -39,7 +41,7 @@ public class AIAgentParallelNodesMergeContext<Input, Output>(
     override val storage: AIAgentStorage get() = underlyingContextBase.storage
     override val runId: String get() = underlyingContextBase.runId
     override val strategyName: String get() = underlyingContextBase.strategyName
-    override val pipeline: AIAgentPipeline<AIAgentGraphStrategy<Input, Output>> get() = underlyingContextBase.pipeline
+    override val pipeline: AIAgentGraphPipeline<*, *> get() = underlyingContextBase.pipeline
 
     override fun store(key: AIAgentStorageKey<*>, value: Any) {
         underlyingContextBase.store(key, value)
@@ -58,14 +60,13 @@ public class AIAgentParallelNodesMergeContext<Input, Output>(
         underlyingContextBase.feature(key)
 
 
-    override fun <Feature : Any> feature(feature: AIAgentFeature<*, Feature, *>): Feature? =
+    override fun <Feature : Any> feature(feature: AIAgentFeature<*, Feature>): Feature? =
         underlyingContextBase.feature(feature)
 
-    override fun <Feature : Any> featureOrThrow(feature: AIAgentFeature<*, Feature, *>): Feature =
+    override fun <Feature : Any> featureOrThrow(feature: AIAgentFeature<*, Feature>): Feature =
         underlyingContextBase.featureOrThrow(feature)
 
     override suspend fun getHistory(): List<Message> = underlyingContextBase.getHistory()
-
     override fun copy(
         environment: AIAgentEnvironment,
         agentInput: Any?,
@@ -75,20 +76,22 @@ public class AIAgentParallelNodesMergeContext<Input, Output>(
         storage: AIAgentStorage,
         runId: String,
         strategyId: String,
-        pipeline: AIAgentPipeline<AIAgentGraphStrategy<Input, Output>>
-    ): AIAgentContextBase<AIAgentGraphStrategy<Input, Output>> = underlyingContextBase.copy(
-        environment = environment,
-        agentInput = agentInput,
-        config = config,
-        llm = llm,
-        stateManager = stateManager,
-        storage = storage,
-        runId = runId,
-        strategyId = strategyId,
-        pipeline = pipeline
-    )
+        pipeline: AIAgentGraphPipeline<*, *>
+    ): AIAgentContextBase<AIAgentGraphPipeline<*, *>> {
+        return underlyingContextBase.copy(
+            environment = environment,
+            agentInput = agentInput,
+            config = config,
+            llm = llm,
+            stateManager = stateManager,
+            storage = storage,
+            runId = runId,
+            strategyId = strategyId,
+            pipeline = pipeline
+        )
+    }
 
-    override suspend fun fork(): AIAgentContextBase<AIAgentGraphStrategy<Input, Output>> = underlyingContextBase.fork()
+    override suspend fun fork(): AIAgentContextBase<AIAgentGraphPipeline<*, *>> = underlyingContextBase.fork()
 
     override suspend fun replace(context: AIAgentContextBase<*>): Unit = underlyingContextBase.replace(context)
