+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/handler/StrategyHandler.kt
@@ -11,7 +11,7 @@ import ai.koog.agents.core.annotation.InternalAgentsApi
  * @param TFeature The type of feature associated with the strategy operations.
  * @property feature The specific feature instance associated with this handler.
  */
-public class StrategyHandler<TFeature : Any, TStrategy: AIAgentStrategy<*, *>>(public val feature: TFeature) {
+public class StrategyHandler<TFeature : Any, TStrategy: AIAgentStrategy<*, *, *>>(public val feature: TFeature) {
 
     /**
      * Handler invoked when a strategy is started. This can be used to perform custom logic
@@ -52,7 +52,7 @@ public class StrategyHandler<TFeature : Any, TStrategy: AIAgentStrategy<*, *>>(p
      */
     @Suppress("UNCHECKED_CAST")
     @InternalAgentsApi
-    public suspend fun handleStrategyStartedUnsafe(context: StrategyStartContext<*, TStrategy>) {
+    public suspend fun handleStrategyStartedUnsafe(context: StrategyStartContext<*, AIAgentStrategy<*, *, *>>) {
         handleStrategyStarted(context as StrategyStartContext<TFeature, TStrategy>)
     }
 
@@ -82,7 +82,7 @@ public class StrategyHandler<TFeature : Any, TStrategy: AIAgentStrategy<*, *>>(p
  *
  * @param FeatureT The type of feature associated with the strategy.
  */
-public fun interface StrategyStartedHandler<FeatureT : Any, TStrategy : AIAgentStrategy<*, *>> {
+public fun interface StrategyStartedHandler<FeatureT : Any, TStrategy : AIAgentStrategy<*, *, *>> {
     /**
      * Handles the processing of a strategy update within a specified context.
      *
