+++ agents/agents-features/agents-features-snapshot/src/jvmTest/kotlin/SubgraphSetExecutionPointTest.kt
@@ -1,4 +1,4 @@
-import ai.koog.agents.core.agent.AIAgent
+import ai.koog.agents.core.agent.agentImpls.AIAgent
 import ai.koog.agents.core.agent.config.AIAgentConfig
 import ai.koog.agents.core.tools.ToolRegistry
 import ai.koog.agents.ext.tool.SayToUser
