+++ examples/src/main/kotlin/ai/koog/agents/example/snapshot/FilePersistentAgentExample.kt
@@ -1,6 +1,6 @@
 package ai.koog.agents.example.snapshot
 
-import ai.koog.agents.core.agent.AIAgent
+import ai.koog.agents.core.agent.agentImpls.AIAgent
 import ai.koog.agents.core.agent.config.AIAgentConfig
 import ai.koog.agents.core.tools.ToolRegistry
 import ai.koog.agents.ext.tool.AskUser
